backend/requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.0.8
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
idna==2.10
oauthlib==3.1.0
# psycopg2==2.8.5
# psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.26.2
blis==0.7.8
cached-property==1.5.2
catalogue==1.0.0
charset-normalizer==2.1.0
cymem==2.0.6
docopt==0.6.2
# en-core-web-sm==2.3.1
en-core-web-sm
idna==3.3
more-itertools==8.13.0
murmurhash==1.0.7
numpy==1.23.0
ordered-set==3.1.1
ordered-set-stubs==0.1.3
Pillow==9.2.0
pipreqs==0.4.11
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.6
profanity-filter==1.3.3
pydantic==1.9.1
redis==3.5.3
requests==2.28.1
ruamel.yaml==0.15.100
spacy==2.3.7
srsly==1.0.5
thinc==7.4.5
tomlkit==0.5.11
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.9
wasabi==0.9.1
websockets==10.3
yarg==0.1.9